Refactor folders
Move component code to separate module, and re-export from index
